requirements.txt
asgiref==3.7.2
Django==4.2.4
Pillow
sqlparse==0.4.4
psycopg2-binary
ipython~=8.14.0
pytils
future~=0.18.2
olefile~=0.46
pip~=23.1.2
wheel~=0.37.1
cryptography~=3.4.8
keyring~=23.5.0
psycopg2~=2.9.6
colorama~=0.4.4
requests~=2.25.1
setuptools~=59.6.0
python-dotenv~=1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
